schedule-risk-backend/requirements.txt
# FastAPI and Server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9

# Environment Variables
python-dotenv>=1.0.0

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
email-validator>=2.0.0

# HTTP Client (for LLM APIs)
httpx>=0.25.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Machine Learning (for ML risk models)
scikit-learn>=1.3.0
joblib>=1.3.0

# LLM Support (Optional)
huggingface-hub>=0.19.0

